tests ongoing

+++ src/main/java/com/climbinggym/controller/PackageController.java
@@ -4,10 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
 import com.climbinggym.entity.Package;
 import com.climbinggym.service.PackageService;
-
 import java.util.List;
 
 @RestController
@@ -26,7 +24,7 @@ public class PackageController {
     public Package createPackage(@RequestBody Package pkg) {
         // Ensure the package availability is set (if not provided, default to true)
         if (pkg.getAvailable() == null) {
-            pkg.setAvailable(true); // Default to true if not provided
+            pkg.setAvailable(true);
         }
         return packageService.addPackage(pkg);
     }
@@ -42,13 +40,13 @@ public class PackageController {
     }
 
     @GetMapping("/{id}")
-    public Package getPackageById(@PathVariable Long id) {
-        // Now directly fetch the package or throw an exception from the service
+    public Package getPackageById(@PathVariable Integer id) {
+        // fetch the package or throw an exception from the service
         return packageService.getPackageById(id);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletePackage(@PathVariable Long id) {
+    public ResponseEntity<Void> deletePackage(@PathVariable Integer id) {
         try {
             packageService.deletePackage(id);
             return ResponseEntity.noContent().build();
@@ -59,7 +57,7 @@ public class PackageController {
 
 
     @PutMapping("/{id}")
-    public Package updatePackage(@PathVariable Long id, @RequestBody Package updatedPackage) {
+    public Package updatePackage(@PathVariable Integer id, @RequestBody Package updatedPackage) {
         // Fetch the existing package or throw an exception from the service
         Package existingPackage = packageService.getPackageById(id);
