+++ src/main/java/com/climbinggym/repository/PaymentRepository.java
@@ -3,11 +3,9 @@ package com.climbinggym.repository;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.climbinggym.entity.Payment;
 
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
 
-    // Define custom query to find payments by user ID
-    List<Payment> findByUserId(Long userId);
+    List<Payment> findByUserId(Integer userId);
 }
