+++ src/main/java/com/climbinggym/repository/UserRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.climbinggym.entity.User;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email); // Example custom query
+public interface UserRepository extends JpaRepository<User, Integer> {
+    Optional<User> findByEmail(String email); 
 
     boolean existsByEmail(String email);
     boolean existsByPhone(String phone);
