+++ src/main/java/com/climbinggym/service/UserService.java
@@ -2,18 +2,18 @@ package com.climbinggym.service;
 
 import java.time.LocalDate;
 import java.util.List;
-import com.climbinggym.entity.Package;  // Ensure this import is present
+import com.climbinggym.entity.Package; 
 import com.climbinggym.entity.User;
 
 public interface UserService {
     User addUser(User user);
-    User getUser(Long id);
+    User getUser(Integer id);
     List<User> getAllUsers();
-    void updateMembership(Long userId, Package purchasedPackage);
-    boolean isMembershipActive(Long id);
-    void deleteUser(Long id);
-    User updateUser(Long id, User updatedUser);
+    void updateMembership(Integer userId, Package purchasedPackage);
+    boolean isMembershipActive(Integer id);
+    void deleteUser(Integer id);
+    User updateUser(Integer id, User updatedUser);
     public boolean isPhoneUnique(String phone);
     public boolean isEmailUnique(String email);
-    public LocalDate getMembershipExpirationDate(Long userId);
+    public LocalDate getMembershipExpirationDate(Integer userId);
 }
